refactor(cdc): split flusher.go batch execution into flush_batch.go (#180)
Pure mechanical move to restore the 500-line file limit: batch execution
mechanics (flushBatchExecutor, executeBatch, buildFlushS3Keys, chunk
drivers, updateManifest, minMaxRowID) move to flush_batch.go; their tests
and the shared test doubles move to flush_batch_test.go and mocks_test.go.
Declaration count is unchanged (69 before/after); no code was modified.
